+++ src/main/java/de/maxhenkel/easypiglins/EasyPiglinsClientMod.java
@@ -2,6 +2,7 @@ package de.maxhenkel.easypiglins;
 
 import de.maxhenkel.easypiglins.blocks.tileentity.ModTileEntities;
 import de.maxhenkel.easypiglins.blocks.tileentity.render.BartererRenderer;
+import de.maxhenkel.easypiglins.events.ModSoundEvents;
 import de.maxhenkel.easypiglins.gui.Containers;
 import de.maxhenkel.easypiglins.items.render.BartererSpecialRenderer;
 import de.maxhenkel.easypiglins.items.render.PiglinSpecialRenderer;
@@ -14,6 +15,7 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.RegisterSpecialModelRendererEvent;
+import net.neoforged.neoforge.common.NeoForge;
 
 @Mod(value = EasyPiglinsMod.MODID, dist = Dist.CLIENT)
 @EventBusSubscriber(modid = EasyPiglinsMod.MODID, value = Dist.CLIENT)
@@ -26,6 +28,8 @@ public class EasyPiglinsClientMod {
     @SubscribeEvent
     static void clientSetup(FMLClientSetupEvent event) {
         BlockEntityRenderers.register(ModTileEntities.BARTERER.get(), c -> new BartererRenderer(c.blockRenderDispatcher()));
+
+        NeoForge.EVENT_BUS.register(new ModSoundEvents());
     }
 
     @SubscribeEvent